% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_ts.R
\name{center_peak}
\alias{center_peak}
\title{Center around the peak value}
\usage{
center_peak(x_mt, f)
}
\arguments{
\item{x_mt}{a numeric matrix. A matrix of cyclic observations; each time
cycle has its own column.}

\item{f}{a character(1). A function for aggregating data across cycles, that
is, by row.}
}
\value{
a data frame with at least 4 columns: the original position of each
observation (pos), the centered position (center_pos), and the translation
necessary to return to the original position (center_trans). The x column is
reordered according to center_pos.
}
\description{
Displaces the rows in the given matrix so that the peak-value row is at the
center position. The peak-value row is found by aggregating the row values
using the given function. The displacement ensures the relative position of
rows is preserved.
}
