% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time.R
\name{compute_period}
\alias{compute_period}
\title{Compute the period of the given date}
\usage{
compute_period(adate, start_month, start_day)
}
\arguments{
\item{adate}{A date vector.}

\item{start_month}{A character(1). The month (number) of the first month of
the period.}

\item{start_day}{A character(1). The day (number) of the first day of the
period.}
}
\value{
An integer.
}
\description{
The period is the same concept as a the year, but it starts on a date
different from January the 1st. The period takes the name of the year on
which the period ends.
}
