% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{compute_prodes_year}
\alias{compute_prodes_year}
\title{Compute the PRODES year of the given date.}
\usage{
compute_prodes_year(adate, start_month = "08", start_day = "01")
}
\arguments{
\item{adate}{A date vector.}

\item{start_month}{A character. The month (number) of the first month of
a PRODES year.}

\item{start_day}{A character. The day (number) of the first day of a PRODES
year.}
}
\value{
An integer.
}
\description{
This function computes the corresponding PRODES year of the given date
(PRODES is the brazilian deforestation monitoring system). The PRODES year
starts on August and it takes the year of its last July.
}
