% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{compute_season}
\alias{compute_season}
\title{Compute the season}
\usage{
compute_season(x, threshold_cons)
}
\arguments{
\item{x}{a numeric. The values observed during a cycle.}

\item{threshold_cons}{a numeric(1) between 0 and 1. The percentage of the
total a season must reach.}
}
\value{
an integer. The positions of the observations in x corresponding to
the season.
}
\description{
Given a vector representing a cycle of observations (e.g. 12 observations),
estimate the peak season. For example, given a year of monthly observations,
the peak season is the minimum subset of consecutive values around the peak
value that add to the threshold of the season total.
}
