% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time.R
\name{date2monthdec}
\alias{date2monthdec}
\title{Convert a date into a decimal month}
\usage{
date2monthdec(adate)
}
\arguments{
\item{adate}{a POSIXlt date}
}
\value{
a list with 2 elements: A year and a decimal month.
}
\description{
Given a date, this function returns its equivalent year and month fraction.
}
