% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering_ts.R
\name{get_cycle_matrix}
\alias{get_cycle_matrix}
\title{Arrange one cycle in each column}
\usage{
get_cycle_matrix(x, n_cycles)
}
\arguments{
\item{x}{a numeric. Data observations.}

\item{n_cycles}{an integer(1). The number of times x cycles over.}
}
\value{
a numeric matrix.
}
\description{
Given a vector of observations and the number of time cycles it covers,
return a matrix where the observations of each cycle are in their own
column.
}
