% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_field_stats}
\alias{get_field_stats}
\title{Compute statistics over the column of a CSV file}
\usage{
get_field_stats(x)
}
\arguments{
\item{x}{A character. Path to a CSV file.}
}
\value{
a data frame with statistics.
}
\description{
Read the given CSV file and compute statistics over its columns.
}
