% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_season.R
\name{get_na_df}
\alias{get_na_df}
\title{Build an empty season data frame}
\usage{
get_na_df()
}
\value{
a data frame.
}
\description{
Create a data frame of NAs with the expected columns of a season data frame.
}
