% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_aggregate.R
\name{ji}
\alias{ji}
\title{Estimate the grid cell of each point.}
\usage{
ji(xy, origin, cellsize)
}
\arguments{
\item{xy}{a matrix with two columns; The X and Y coordinates.}

\item{origin}{a numeric(2). The origin of the grid.}

\item{cellsize}{a numeric(2). The size of a cell in the grid.}
}
\value{
a numeric matrix with 2 columns representing the longitude and
latitude of thecenter point of the cell corresponding to each point in the
given data frame.
}
\description{
Estimate the grid cell of each point. Taken from whuber's answer at
}
