% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_time.R
\name{monthdec2date}
\alias{monthdec2date}
\title{Convert from decimal month to date}
\usage{
monthdec2date(year, month_dec, time_zone = "UTC")
}
\arguments{
\item{year}{an integer. A number representing a year.}

\item{month_dec}{a numeric. A number (1 <= x < 13) representing a month.}

\item{time_zone}{a character(1). The time zone of the resulting date.}
}
\value{
a POSIXct or POSIXt object. The date corresponding to the input data.
}
\description{
Given a decimal month, return the equivalent date.
}
