% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.R
\name{rasterize_points}
\alias{rasterize_points}
\title{Rasterize points}
\usage{
rasterize_points(vname, data_sf, grid_r)
}
\arguments{
\item{vname}{a character. Name of a column in data_sf.}

\item{data_sf}{an sf object (points).}

\item{grid_r}{a terra object (raster).}
}
\value{
a terra object (raster).
}
\description{
Rasterize points given a variable name. This is a utility function.
}
