% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_peak-threshold.R
\name{season_peak_thres_helper}
\alias{season_peak_thres_helper}
\title{Compute the peak seaon using a threshold}
\usage{
season_peak_thres_helper(
  x,
  id_group,
  id_col,
  val_col,
  month_col,
  threshold_cons
)
}
\arguments{
\item{x}{a data frame.}

\item{id_group}{a character. Name of the group column in x.}

\item{id_col}{a character. Name of the ID column in x.}

\item{val_col}{a character. Name of the column with a value.}

\item{month_col}{a character. Name of the column with month numbers.}

\item{threshold_cons}{a numeric(1) between 0 and 1. The percentage of the
total a season must reach. See \link{compute_season_peak_threshold}.}
}
\value{
a data frame (tibble).
}
\description{
Utility funcion for computing the peak threshold.
}
