% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{sort_cols}
\alias{sort_cols}
\title{Sort columns by name}
\usage{
sort_cols(df)
}
\arguments{
\item{df}{a data frame.}
}
\value{
a data frame.
}
\description{
Sort the columns by name in a data frame.
}
