% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{blank_raster}
\alias{blank_raster}
\title{Raster template for rasterization}
\usage{
blank_raster(grid_cells, xy_min, xy_max, grid_crs)
}
\arguments{
\item{grid_cells}{a numeric. Number of cells for the raster.}

\item{xy_min, xy_max}{a numeric. Minimum and maximum XY coordinates.}

\item{grid_crs}{a numeric. EPSG code.}
}
\value{
a terra object (raster).
}
\description{
Raster template for rasterization
}
