% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{center_peak}
\alias{center_peak}
\title{Center around the peak value}
\usage{
center_peak(x)
}
\arguments{
\item{x}{a numeric. A vector of cyclic observations.}
}
\value{
a data frame with 3 columns: x, the original position of each
observation (pos), and the centered position (center_pos). The x column is
reordered according to center_pos.
}
\description{
Center the given vector around its peak (maximum value). In this way, the
peak would be at the center position in the vector.
}
