% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{compute_season_peak_threshold}
\alias{compute_season_peak_threshold}
\title{Compute the season using its peak and a threshold}
\usage{
compute_season_peak_threshold(x, threshold_cons)
}
\arguments{
\item{x}{a numeric. The values observed during a cycle.}

\item{threshold_cons}{a numeric(1) between 0 and 1. The percentage of the
total a season must reach.}
}
\value{
a data frame with season metrics.
}
\description{
Given a vector representing a cycle of observations (e.g. 12 observations),
estimate the season start and end using the season peak and a threshold. For
example, given a year of monthly observations, the peak season is the
minimum subset of consecutive values around the peak value that reach the
a given threshold.
}
