% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{compute_season_peak_threshold}
\alias{compute_season_peak_threshold}
\alias{compute_season_double_sig}
\title{Compute a season using cyclical observations}
\usage{
compute_season_peak_threshold(x, threshold_cons, x_cycles = 1L, f = "sum")

compute_season_double_sig(x, x_cycles = 1, n_runs_min = 20, n_runs_max = 500)
}
\arguments{
\item{x}{a numeric. A vector of cyclical observations.}

\item{threshold_cons}{a numeric(1) between 0 and 1. The percentage of the
total a season must reach.}

\item{x_cycles}{a integer(1). Number of cycles in x.}

\item{f}{a character(1). A function to apply to aggregate x along cycles.}

\item{n_runs_min, n_runs_max}{an integer(1). Minimum and maximum number of
successful fitting attempts.}
}
\value{
a data frame with metrics:
\itemize{
\item \code{pos_from} and \code{pos_to} are the indices of the season's first and last
elements in \code{x}.
\item \code{val_from} and \code{val_to} area the values at \code{pos_from} and \code{pos_to}.
\item \code{pos_min} and \code{pos_max} are the indices where the season reaches its
mininum and maximum values.
\item \code{val_min} and \code{val_max} are the values at \code{pos_min} and \code{pos_max}.
\item \code{val_len} is the season length. Note that the first element is included
when estimating the length for \code{compute_season_peak_threshold}.
\item \code{val_mean} is the mean of the values in the season.
\item \code{val_sd} is the standard deviation of the values in the season.
}
}
\description{
Cyclical observations refers to vectors on which the element before the
first is the last one, and the element after the last vector element is the
first one.

\code{compute_season_peak_threshold} estimate season's parameters using the
season maximum value and a threshold. For example, given a year of monthly
observations, the peak season is the minimum subset of consecutive values
around the maximum value (the peak) that reach the given threshold.

\code{compute_season_double_sig} estimate season's parameters by adjusting a
double sigmoidal function to the observations.
}
