% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{displace_vec}
\alias{displace_vec}
\title{Displace vector elements}
\usage{
displace_vec(x, n_pos)
}
\arguments{
\item{x}{A vector.}

\item{n_pos}{an integer. Number of positions to displace the vector's
elements.}
}
\value{
the given vector with elements in a different order.
}
\description{
Displace the elements in the given vector (to the left) by certain number of
positions.
}
