% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{get_prev_next}
\alias{get_prev_next}
\title{Determine the next values for computing season}
\usage{
get_prev_next(y, total_len)
}
\arguments{
\item{y}{an integer vector with the current season (positions of the season
values).}

\item{total_len}{an integer(1). The total number of elements is a cycle.}
}
\value{
an integer(2) with the month before and after the given season.
}
\description{
Utility function. Get the next values to evaluate.
}
