% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dsig.R
\name{season_peak_dsig_helper}
\alias{season_peak_dsig_helper}
\title{Compute the peak seaon using a double sigmoidal function}
\usage{
season_peak_dsig_helper(
  x,
  id_group,
  id_col,
  val_col,
  month_col,
  n_runs_min,
  n_runs_max
)
}
\arguments{
\item{x}{a data frame.}

\item{id_group}{a character. Name of the group column in x.}

\item{id_col}{a character. Name of the ID column in x.}

\item{val_col}{a character. Name of the column with a value.}

\item{month_col}{a character. Name of the column with month numbers.}

\item{n_runs_min, n_runs_max}{an integer(1). Minimum and maximum number of
successful fitting attempts. See \link{compute_season_double_sig}.}
}
\value{
a data frame (tibble).
}
\description{
Utility funcion for computing the peak season using a double sigmoidal
function fit.
}
