% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{un_center}
\alias{un_center}
\title{Uncentering an observation position}
\usage{
un_center(pos, x_df)
}
\arguments{
\item{pos}{a numeric(1). A centered position of an observation in a vector.}

\item{x_df}{a data frame. This data frame contains columns correspondign to
observations (x), their original positions (pos), their positions centered
(center_pos) and the translation requited to return the centered positions
to their original place (center_trans).}
}
\value{
a numeric. The transformation constant to return the given centered
position to its original place.
}
\description{
This function undoes the effects of \code{center_peak} by compensating the given
position to its original place.
}
